experiments-avih-baseline
because the patches at this branch are always rebased on top of mpv-master,
which happens much more frequently than adding patches, the HEAD hash doesn't
mean much and doesn't really help tracking the code for this build.

so use the mpv-git-master revision over which the patches are rebased, with
the addition of "+N@avih=<HEAD>" where N counts the additional patches.

this still doesn't help tracking which set of patches are applied on top
because <HEAD> change every rebase, but it's better than just <HEAD>.