js: version.sh: add (master-<mpv master rev> +JS)
because the js patches are always rebased on top of mpv-master, which happens
much more frequently than adding js patches, the HEAD hash doesn't mean much
and doesn't really help tracking the code for this build.

so add the mpv-git-master revision over which the js patches are rebased, with
the addition of "+JS".

this still doesn't help tracking which set of js patches are applied on top, but
it's better than before.