command: new property: mouse-pos
This is a read-only MPV_NODE value with fields: x, y.
The values are unmodified from mp_input_get_mouse_pos(...).

For now the property doesn't trigger observer notifications, because
it's not always associated with an event, or maybe I'm just missing
some obvious way to do that.

If it can be made observe-able, then this could be used for global
mouse-move (and in the future enter/leave) triggers such that more than
one client can be notified (while now only the topmost mouse-area
binding is triggered).

Also make the undocumentd mp.get_mouse_pos() (lua, js) a wrapper for
this property.