command: mouse-pos property: add field "hover"
Add a third field: "hover", which is updated from input.c after input
keys MP_KEY_MOUSE_LEAVE and MP_KEY_MOUSE_ENTER - which are typically
sent by the VO.

Previously an arbitrary client could not check if the mouse hovers the
window because the mouse-leave event was only send to the bound
mouse-area which is typically the osc, and the mouse position does not
update when the pointer is outside the window.

The lua mp.get_mouse_pos() wrapper still returns only x, y because
that's what osc.lua needs. Other clients can simply read the property.