command: mouse-pos property: notify observers
This requires input.c to use MPContext, which it didn't till now. Also
the property update code seems to do relatively much work and locking,
so I'm not sure if there's a way to make it more efficient (possibly
would be useful coalesc mouse moves, though not sure when to decide it's
enough).