js: support UTF-8 mujs API with a header-only wrapper
The header provides wrappers to all the mujs string APIs and replaces
the default names with the wrapped names such that nothing is required
other than including it after mujs.h .

The result is that UTF-8 strings are automatically converted to CESU-8
(if required) when sending them to mujs, and returned strings from mujs
are also automatically converted from CESU-8 to UTF-8 (if required).

The wrapper header - mujsutf8.h - is stand alone and does not depend
on any mpv code or headers. It's still WIP (though should be fully
working and correct), and still missing the option to be used as
header/implementation (currently it supports only header-only), as well
as other refinements.